I am attempting to mess around with the turning stuff in order to make the rotate the way Riley wants it. I wanted to keep our previous code safe, so I created a copy of the MecanumLinearOpMode named OldLinearOpMode and made copies of the autos.
I put angles = get angular oreination in the init() method and got rid of updatevalues so it wouldn't keep resetting it. I also added an offset variable right after waitforstart in the auto so that it would store exactly how much offset angle there is from where we intialized it to where we started hanging, so it would add that value to all our turns.

-Asha